image_converter/converter/requirements.txt
Pillow==10.3.0
argparse==1.4.0